feat: custom collection card template
